Update and rename akchilov3消灭一切打表式反射.cpp to akchilov3 第十题.cpp
https://gcc.godbolt.org/z/b63qv93xb
